Fix part of #10415: Refactor prod validators to be split into individual validators (#11493)
